requirements_minimal.txt
# Dependências mínimas para o projeto
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
prophet>=1.1.0
xgboost>=2.0.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0
fastparquet>=2023.0.0

